Fixes #1397 - Adds the delete button back to the history recyclerview
