+++ app/src/main/res/layout/component_history.xml
@@ -9,30 +9,6 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical">
-    <FrameLayout
-        android:id="@+id/delete_history_button"
-        android:foreground="?android:attr/selectableItemBackground"
-        android:background="@drawable/button_background"
-        android:layout_margin="16dp"
-        android:padding="6dp"
-        android:clickable="true"
-        android:focusable="true"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <TextView
-            android:id="@+id/delete_history_button_text"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="@string/history_delete_all"
-            android:textColor="?attr/deleteColor"
-            android:drawablePadding="8dp"
-            android:textSize="16sp"
-            android:gravity="center"
-            android:clickable="false"
-            android:focusable="false"
-            android:layout_gravity="center" />
-    </FrameLayout>
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/history_list"
         android:layout_width="match_parent"
