+++ app/src/main/res/layout/delete_history_button.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- This Source Code Form is subject to the terms of the Mozilla Public
+   - License, v. 2.0. If a copy of the MPL was not distributed with this
+   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
+<FrameLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/delete_history_button"
+    android:foreground="?android:attr/selectableItemBackground"
+    android:background="@drawable/button_background"
+    android:layout_margin="16dp"
+    android:padding="6dp"
+    android:clickable="true"
+    android:focusable="true"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content">
+
+    <TextView
+        android:id="@+id/delete_history_button_text"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="@string/history_delete_all"
+        android:textColor="?attr/deleteColor"
+        android:drawablePadding="8dp"
+        android:textSize="16sp"
+        android:gravity="center"
+        android:clickable="false"
+        android:focusable="false"
+        android:layout_gravity="center" />
+</FrameLayout>
